Remove the elasticsearch plugin on models that are inherited from GoodsNomenclature
There is an issue when you try and recreate your DB from scratch where
delete calls will be made to documents not in ES that will cause the
whole sync process to crash. This is because this plugin uses into the
after_destroy callback.

This removes the sequel plugin from models which are already indexed as part of
the sync process `TradeTariffBackend.reindex`.